UML Diagrams In Vim
Using the pluging vim-slumlord allows the creation of uml diagrams in
vim on save.
The Diagrams are ASCII art and require PlantUML's jar file to be in your
computer's path. I have included it in the misk folder and added it to
vim's path.
You can also use PlantUML to create images of the diagram defined by the
file.